ts electron

+++ packages/suite-desktop/src/bridge.ts
@@ -1,13 +1,13 @@
-const psList = require('ps-list');
-const os = require('os');
-const isDev = require('electron-is-dev');
-const { join } = require('path');
-const { spawn, exec } = require('child_process');
+import psList from 'ps-list';
+import * as os from 'os';
+import isDev from 'electron-is-dev';
+import { join } from 'path';
+import { spawn, exec } from 'child_process';
 
 const TREZOR_PROCESS_NAME = 'trezord';
 const STATUS = { OK: 'ok', ERROR: 'error' };
 
-const error = msg => {
+const error = (msg: string | Error) => {
     throw new Error(`cannot run bridge library - ${msg}`);
 };
 
@@ -66,7 +66,7 @@ const getBridgeLibByOs = () => {
     }
 };
 
-const spawnProcess = command => {
+const spawnProcess = (command: string) => {
     const spawnedProcess = spawn(command, [], {
         detached: true,
     });
@@ -75,13 +75,13 @@ const spawnProcess = command => {
     });
 };
 
-const execute = command => {
-    exec(command, (error, stdout, stderr) => {
+const execute = (command: string) => {
+    exec(command, (error, _stdout, stderr) => {
         if (error) {
-            error(error);
+            console.error(error);
             return;
         }
-        error(`${stderr}`);
+        console.log(`${stderr}`);
     });
 };
 
@@ -92,8 +92,6 @@ const isBridgeRunning = async () => {
 };
 
 const runBridgeProcess = async () => {
-    const os = getOS();
-    const arch = getArch();
     const isBridgeAlreadyRunning = await isBridgeRunning();
 
     // bridge is already installed and running, nothing to do
@@ -102,13 +100,17 @@ const runBridgeProcess = async () => {
     }
 
     const lib = getBridgeLibByOs();
-    spawnProcess(lib);
+    if (lib) {
+        spawnProcess(lib);
+    }
 };
 
 const killBridgeProcess = async () => {
     const processes = await psList();
     const bridgeProcess = processes.find(ps => ps.name.includes(TREZOR_PROCESS_NAME));
-    execute(`kill -9 ${bridgeProcess.pid}`);
+    if (bridgeProcess) {
+        execute(`kill -9 ${bridgeProcess.pid}`);
+    }
 };
 
 module.exports = {
